requirements.txt
Django==5.2.8
djangorestframework==3.16.1
django-redis==5.4.0
django-environ==0.11.2
django-cors-headers==4.3.1
requests==2.31.0
redis==5.0.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0

